+++ src/client/index.ts
@@ -2,7 +2,7 @@ import { Game } from "@client/game";
 import { View } from "@client/view";
 import io from "socket.io-client";
 
-const DEBUG_MODE = false;
+const DEBUG_MODE = true;
 
 // Create socket.io connection to server
 const socket = io();