+++ src/server/cluster-gen/biome-gen/ContinuousBiomeGenerator.ts
@@ -2,6 +2,7 @@ import { Biome } from "@share/utility/Biome";
 import { IContinuousBiomeGenerator } from "./IContinuousBiomeGenerator";
 import { BiomeComposition, IBiomeComposition } from "./biome-composition";
 import { logistic } from "../../../share/utility/Logistic";
+import { BiomePercentage } from "./biome-composition/biome-percentage/BiomePercentage";
 
 /**
  * Class for biome placement generator