+++ src/server/cluster-gen/hill-gen/biome-dependent/BiomeDependentHillGenerator.ts
@@ -4,6 +4,7 @@ import { a } from "@share/data/coordinate/CoordinateGenerators";
 import { logistic } from "@share/utility/Logistic";
 import { HillProps } from "./HillProps";
 import { IContinuousBiomeGenerator } from "@server/cluster-gen/biome-gen";
+import { BiomeHillPropData } from "./HillProps/BiomeHillPropData";
 
 /**
  * This hill generator considers the biome when determining hill properties.
@@ -40,10 +41,11 @@ export class BiomeDependentHillGenerator implements IHillGenerator {
    */
   getYFromXZ(x: number, z: number): number {
 
-    const biomeComposition = this.biomeGen.getBiomesFromXZ(x, z);
-    const hillProps = HillProps.createFrom(biomeComposition);
+    //const biomeComposition = this.biomeGen.getBiomesFromXZ(x, z);
+    //const hillProps = HillProps.createFrom(biomeComposition);
+    const hillProps = BiomeHillPropData.get(this.biomeGen.getTopBiomeFromXZ(x, z));
 
-    const originY = Math.floor(hillProps.getAltitude() + (hillProps.getHeight() / 2));
+    const originY = Math.round(hillProps.getAltitude() + (hillProps.getHeight() / 2));
     const origin = new AbsoluteCoordinate(x, originY, z);
 
     // Determine how high this column should be based on surrounding particles
@@ -68,9 +70,9 @@ export class BiomeDependentHillGenerator implements IHillGenerator {
     if (influenceSum == 0) {
       y = 0;
     } else {
-      y = Math.floor(ySum / influenceSum);
+      y = Math.round(ySum / influenceSum);
     }
 
-    return Math.floor(y);
+    return y;
   }
 }