+++ src/share/utility/index.ts
@@ -3,3 +3,4 @@ export {Block} from "./Block";
 export {Face} from "./Face";
 export {FaceVectorConverter} from "./FaceVectorConverter";
 export {TextureUvCalculator} from "./uvs/TextureUvCalculator";
+export {getBlockFromBiome} from "./getBlockFromBiome";