+++ changelog.md
@@ -1,5 +1,10 @@
 ## 0.0.3
 
+* Add complex terrain generation with logistic curves
 * Add interpolation between biomes
 
 ## 0.0.2
+
+* First version.
+* Sinusoidal terrain generation
+* Block placing / breaking