+++ changelog.md
@@ -1,6 +1,7 @@
 ## 0.0.4
 
 * Fixed lines between blocks
+* Improved top-down biome generation (inverse instead of logistic)
 * PLANNED: Sound FX
 * PLANNED: Fix player movement behavior (going under arches, placing blocks under self, etc.)
 * PLANNED: Make asset manager static