+++ changelog.md
@@ -1,10 +1,13 @@
+## Backlog
+* Sound FX
+* Fix player movement behavior (going under arches, placing blocks under self, etc.)
+* Make asset manager static
+
 ## 0.0.4
 
 * Fixed lines between blocks
 * Improved top-down biome generation (inverse instead of logistic)
-* PLANNED: Sound FX
-* PLANNED: Fix player movement behavior (going under arches, placing blocks under self, etc.)
-* PLANNED: Make asset manager static
+
 
 ## 0.0.3